+++ src/main/java/org/alliancegenome/curation_api/model/entities/VocabularyTerm.java
@@ -17,6 +17,8 @@ import org.alliancegenome.curation_api.model.entities.base.AuditedObject;
 import org.alliancegenome.curation_api.model.entities.base.GeneratedAuditedObject;
 import org.alliancegenome.curation_api.view.View;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
+import org.hibernate.annotations.Fetch;
+import org.hibernate.annotations.FetchMode;
 import org.hibernate.envers.Audited;
 import org.hibernate.search.engine.backend.types.Aggregable;
 import org.hibernate.search.engine.backend.types.Searchable;
@@ -40,8 +42,12 @@ import lombok.ToString;
 @Data
 @EqualsAndHashCode(onlyExplicitlyIncluded = true, callSuper = true)
 @ToString(exclude = { "vocabulary", "vocabularyTermSets" })
-@Table(indexes = { @Index(name = "vocabularyterm_name_index", columnList = "name"),
-		@Index(name = "vocabularyterm_vocabulary_id_index", columnList = "vocabulary_id")})
+@Table(indexes = { 
+	@Index(name = "vocabularyterm_name_index", columnList = "name"),
+	@Index(name = "vocabularyterm_vocabulary_id_index", columnList = "vocabulary_id"),
+	@Index(name = "vocabularyterm_createdby_index", columnList = "createdBy_id"), 
+	@Index(name = "vocabularyterm_updatedby_index", columnList = "updatedBy_id"),
+})
 @Schema(name = "VocabularyTerm", description = "POJO that represents the Vocabulary Term")
 @AGRCurationSchemaVersion(min = "1.2.0", max = LinkMLSchemaConstants.LATEST_RELEASE, dependencies = { AuditedObject.class })
 public class VocabularyTerm extends GeneratedAuditedObject {
@@ -66,12 +72,14 @@ public class VocabularyTerm extends GeneratedAuditedObject {
 	@ManyToMany
 	@JoinTable(indexes = { @Index(columnList = "vocabularyterm_id"), @Index(columnList = "crossreferences_id") })
 	@JsonView({ View.FieldsAndLists.class })
+	@Fetch(FetchMode.SUBSELECT)
 	private List<CrossReference> crossReferences;
 
 	@IndexedEmbedded(includeDepth = 1)
 	@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
 	@ManyToOne
 	@JsonView({ View.VocabularyTermView.class, View.VocabularyTermUpdate.class, View.VocabularyTermSetView.class })
+	@Fetch(FetchMode.JOIN)
 	private Vocabulary vocabulary;
 
 	@FullTextField(analyzer = "autocompleteAnalyzer", searchAnalyzer = "autocompleteSearchAnalyzer")
@@ -86,5 +94,6 @@ public class VocabularyTerm extends GeneratedAuditedObject {
 	@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
 	@ManyToMany(mappedBy = "memberTerms")
 	@JsonView({ View.VocabularyTermView.class, View.VocabularyTermUpdate.class })
+	@Fetch(FetchMode.SUBSELECT)
 	private List<VocabularyTermSet> vocabularyTermSets;
 }