+++ src/main/java/org/alliancegenome/curation_api/model/entities/ontology/OntologyTerm.java
@@ -22,6 +22,8 @@ import org.alliancegenome.curation_api.model.entities.Synonym;
 import org.alliancegenome.curation_api.model.entities.base.AuditedObject;
 import org.alliancegenome.curation_api.model.entities.base.CurieAuditedObject;
 import org.alliancegenome.curation_api.view.View;
+import org.hibernate.annotations.Fetch;
+import org.hibernate.annotations.FetchMode;
 import org.hibernate.envers.Audited;
 import org.hibernate.search.engine.backend.types.Aggregable;
 import org.hibernate.search.engine.backend.types.Searchable;
@@ -100,6 +102,7 @@ public class OntologyTerm extends CurieAuditedObject {
 	@ManyToMany
 	@JoinTable(indexes = @Index(columnList = "ontologyterm_curie", name = "ontologyterm_synonym_ontologyterm_curie_index"))
 	@JsonView({ View.FieldsAndLists.class })
+	@Fetch(FetchMode.SUBSELECT)
 	private List<Synonym> synonyms;
 
 	@IndexedEmbedded(includeDepth = 1)
@@ -108,22 +111,27 @@ public class OntologyTerm extends CurieAuditedObject {
 	@org.hibernate.annotations.OnDelete(action = org.hibernate.annotations.OnDeleteAction.CASCADE)
 	@JoinTable(indexes = { @Index(columnList = "ontologyterm_curie", name = "ontologyterm_crossreference_ontologyterm_curie_index"), @Index(columnList = "crossreferences_id", name = "ontologyterm_crossreference_crossreferences_id_index") })
 	@JsonView({ View.FieldsAndLists.class })
+	@Fetch(FetchMode.SUBSELECT)
 	private List<CrossReference> crossReferences;
 
 	@ManyToMany
 	// @JsonView(View.OntologyTermView.class)
 	@JoinTable(name = "ontologyterm_isa_parent_children", indexes = { @Index(columnList = "isaparents_curie"), @Index(columnList = "isachildren_curie") })
+	@Fetch(FetchMode.SUBSELECT)
 	private Set<OntologyTerm> isaParents;
 
 	@ManyToMany(mappedBy = "isaParents")
 	// @JsonView(View.OntologyTermView.class)
+	@Fetch(FetchMode.SUBSELECT)
 	private Set<OntologyTerm> isaChildren;
 
 	@ManyToMany
 	@JoinTable(name = "ontologyterm_isa_ancestor_descendant", indexes = { @Index(columnList = "isaancestors_curie"), @Index(columnList = "isadescendants_curie") })
+	@Fetch(FetchMode.SUBSELECT)
 	private Set<OntologyTerm> isaAncestors;
 
 	@ManyToMany(mappedBy = "isaAncestors")
+	@Fetch(FetchMode.SUBSELECT)
 	private Set<OntologyTerm> isaDescendants;
 
 	@Transient