+++ src/main/java/org/alliancegenome/curation_api/model/entities/slotAnnotations/SlotAnnotation.java
@@ -17,6 +17,8 @@ import org.alliancegenome.curation_api.model.entities.base.AuditedObject;
 import org.alliancegenome.curation_api.model.entities.base.GeneratedAuditedObject;
 import org.alliancegenome.curation_api.view.View;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
+import org.hibernate.annotations.Fetch;
+import org.hibernate.annotations.FetchMode;
 import org.hibernate.envers.Audited;
 import org.hibernate.search.mapper.pojo.automaticindexing.ReindexOnUpdate;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
@@ -47,6 +49,7 @@ public class SlotAnnotation extends GeneratedAuditedObject {
 
 	})
 	@JsonView({ View.FieldsAndLists.class, View.AlleleView.class, View.GeneView.class })
+	@Fetch(FetchMode.SUBSELECT)
 	private List<InformationContentEntity> evidence;
 
 }